+++ src/test/kotlin/com/zepben/ewb/services/network/tracing/networktrace/actions/EquipmentTreeBuilderTest.kt
@@ -25,6 +25,7 @@ import io.mockk.verify
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.*
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.api.extension.RegisterExtension
 import java.util.*
 
@@ -32,6 +33,44 @@ internal class EquipmentTreeBuilderTest {
     @JvmField
     @RegisterExtension
     val systemErr: SystemLogExtension = SystemLogExtension.SYSTEM_ERR.captureLog().muteOnSuccess()
+
+    @Test
+    fun `accessing leaves when not calculated throws exception`() {
+        val builder = EquipmentTreeBuilder()
+        assertThrows<IllegalArgumentException> { builder.leaves }
+    }
+
+    @Test
+    fun `test equipment tree builder leaves`() {
+        val n = LoopingNetwork.create()
+
+        n.get<ConductingEquipment>("j0")!!.addFeederDirections().also { DirectionLogger.trace(it) }
+
+        val treeBuilder = EquipmentTreeBuilder(calculateLeaves = true)
+        Tracing.networkTraceBranching()
+            .addCondition { downstream() }
+            .addStepAction(treeBuilder)
+            .run(n.get<ConductingEquipment>("j1")!!)
+
+        assertThat(
+            treeBuilder.leaves.map { leaf -> leaf.identifiedObject.mRID },
+            contains(
+                "j6",
+                "j5",
+                "j6",
+                "acLineSegment16",
+                "acLineSegment12",
+                "j13",
+                "acLineSegment11",
+                "j13",
+                "acLineSegment15",
+                "acLineSegment12",
+                "acLineSegment11",
+                "j13"
+            )
+        )
+
+    }
 
     @Test
     fun `computes initial value`() {